Motor and pressure sensor tests.
The motor works by applying a constant 2v voltage to the ESC and keeping it connected. This is then adjusted between 2 and 5 v for no to full throttle.

The pressure sensor works but connecting the pressure sensor to the power and the A4 and A5 pins for SCL and SDA